python-agents/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
supabase==2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1